input: atmel_mxt_ts_mmi: support for multiple active/suspend configs
Recently added features require different settings applied to multiple
objects depending on the currently effective state: active or suspended.
We found a concept of run time patch sets to be quite useful for that.
Patch set is a number of changes to one or more registers of one or more
objects applied at once. Patch sets come in pairs, where one set
corresponds to active state and the other to suspended. This concept is
extended by adding 2 different modes: default and alternative. Mode
corresponds to certain touch behavior. Driver switches between modes
when a trigger is received. In this change supported trigger is a
touch sequence processor request sent via sysfs.

Change-Id: I64e7c599d20fa06a02257b6ad30d724f7c4399b1
Signed-off-by: Konstantin Makariev <kmakariev@motorola.com>